refactor(patterns): split oversized modules into focused helpers
