+++ dlt/google_sheets/helpers/data_processing.py
@@ -1,13 +1,15 @@
 """This is a helper module that contains function which validate and process data"""
 
 import re
-from typing import Any, Iterator, List, Tuple, Union, NamedTuple, Optional
+from typing import Any, Iterator, List, Union, Optional
 
 import dlt
 from dlt.common import logger, pendulum
 from dlt.common.typing import DictStrAny, TTableHintTemplate
 from dlt.common.data_types import TDataType
 from dlt.common.schema.typing import TTableSchemaColumns
+
+from .range_utils import ParsedRange, trim_range_top_left
 
 # this string comes before the id
 URL_ID_IDENTIFIER = "d"
@@ -17,65 +19,17 @@ SECONDS_PER_DAY = 86400
 DLT_TIMEZONE = "UTC"
 # number of seconds from UNIX timestamp origin (1st Jan 1970) to serial number origin (30th Dec 1899)
 TIMESTAMP_CONST = -2209161600.0
-# compiled regex to extract ranges
-RE_PARSE_RANGE = re.compile(
-    r"^(?:(?P<sheet>[\'\w\s\-]+)!)?(?P<start_col>[A-Z]+)(?P<start_row>\d+):(?P<end_col>[A-Z]+)(?P<end_row>\d+)$"
-)
 
-
-class ParsedRange(NamedTuple):
-    sheet_name: str
-    start_col: str
-    start_row: int
-    end_col: str
-    end_row: int
-
-    @classmethod
-    def parse_range(cls, s: str) -> "ParsedRange":
-        match = RE_PARSE_RANGE.match(s)
-        if match:
-            parsed_dict = match.groupdict()
-            return ParsedRange(
-                parsed_dict["sheet"].strip("'"),
-                parsed_dict["start_col"],
-                int(parsed_dict["start_row"]),
-                parsed_dict["end_col"],
-                int(parsed_dict["end_row"]),
-            )
-        else:
-            raise ValueError(s)
-
-    def __str__(self) -> str:
-        return f"{self.sheet_name}!{self.start_col}{self.start_row}:{self.end_col}{self.end_row}"
-
-    @staticmethod
-    def shift_column(col: str, shift: int) -> str:
-        """
-        Shift a Google Sheets column string by a given number of positions.
-
-        Parameters:
-        col (str): The original column string.
-        shift (int): The number of positions to shift the column.
-
-        Returns:
-        str: The new column string after shifting.
-        """
-        # Convert column string to column index (1-indexed)
-        col_num = 0
-        for i, char in enumerate(reversed(col)):
-            col_num += (ord(char.upper()) - 65 + 1) * (26**i)
-
-        # Shift the column index
-        col_num += shift
-
-        # Convert back to column string
-        col_str = ""
-        while col_num > 0:
-            col_num, remainder = divmod(col_num - 1, 26)
-            col_str = chr(65 + remainder) + col_str
-
-        return col_str
-
+__all__ = [
+    "ParsedRange",
+    "trim_range_top_left",
+    "get_spreadsheet_id",
+    "extract_spreadsheet_id_from_url",
+    "get_range_headers",
+    "get_data_types",
+    "serial_date_to_datetime",
+    "process_range",
+]
 
 def get_spreadsheet_id(url_or_id: str) -> str:
     """
@@ -317,33 +271,3 @@ def process_range(
         yield table_dict
 
 
-def trim_range_top_left(
-    parsed_range: ParsedRange, range_values: List[List[Any]]
-) -> Tuple[ParsedRange, List[List[Any]]]:
-    # skip empty rows and then empty columns
-    # skip empty rows
-    shift_x = 0
-    for row in range_values:
-        if row:
-            break
-        else:
-            shift_x += 1
-    if shift_x > 0:
-        range_values = range_values[shift_x:]
-    # skip empty columns
-    shift_y = 0
-    if len(range_values) > 0:
-        for col in range_values[0]:
-            if col == "":
-                shift_y += 1
-            else:
-                break
-        if shift_y > 0:
-            # skip all columns
-            for idx, row in enumerate(range_values):
-                range_values[idx] = row[shift_y:]
-    parsed_range = parsed_range._replace(
-        start_row=parsed_range.start_row + shift_x,
-        start_col=ParsedRange.shift_column(parsed_range.start_col, shift_y),
-    )
-    return parsed_range, range_values
